build: replace commitizen hook with cm script to open the CLI
`prepare-commit-msg` hook was being used to open the Commitizen CLI,
but it caused the CD process to fail when a new version had to be
released, and it was also being triggered in merge commits. For those
reasons it is being replaced with the `cm` script.
Commitlint is still used both at client and at CI level, so the
removal of the hook is for better, as the CLI may be annoying for
those already familiar with the commit message convention.